requirements.txt
amqp==5.2.0
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-SocketIO==5.3.6
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
humanize==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
packaging==24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
